+++ examples/data-stores/bucket-home-folders/variables.tf
@@ -11,7 +11,7 @@ variable "bucket_name" {
 
 variable "home_folder_names" {
   description = "The list of home folder names"
-  type        = list(string)
+  type        = set(string)
   default = ["myfirstfolder", "mysecondfolder"]
 }