+++ modules/data-stores/bucket-home-folders/main.tf
@@ -1,21 +1,12 @@
-terraform {
-  required_version = ">= 0.12, < 0.13"
-}
-
-provider "aws" {
-  region = "eu-central-1"
-  version = "~> 2.0"
-}
-
 data "aws_s3_bucket" "bucket" {
   bucket = var.bucket_name
 }
 
 resource "aws_s3_bucket_object" "home_folders" {
-  count = var.create_personal_folders? length(var.home_folder_names) : 0
+  for_each = var.home_folder_names
   
   bucket = data.aws_s3_bucket.bucket.id
-  key    = format("%s/", element(var.home_folder_names, count.index))
+  key    = format("%s/", each.value)
   source = "/dev/null"
 }