+++ test/go_sanity_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 // GoIsWorking performs a sanity check
-func GoIsWorking(t *testing.T) {
+func TestGoSanity(t *testing.T) {
 	fmt.Println()
 	fmt.Println("If you see this, it is working!")
 	fmt.Println()