+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+* Update module to Terraform ~> 1.2.6
+* Switch from `list` to `set` for `user_names` and `home_folder_names` variables
+
 ## 1.2.0
 
 * S3 bucket is no longer created by this module